added function
added render_animation_with_video_mask
allows a mask video to be used when video init is not wanted. 
The purpose is to be able to apply an animated masks when not using init video.
Work can be done to delete the old directory and replace it. Currently you have to delete mask frames independently. That will be for another session of fixes however.