Fixed issue where depth only worked properly at 1:1 aspect ratio
This has been an issue with Deforum as far back as I can remember... If the proportion was too wide, things would stretch unnaturally vertically, and if proportion was too thin, things would stretch horizontally.

I found that changing the aspect ratio in this function to 1, things are always correct, no matter the proportion of the image!

This is a major change to Deforum, but in all of my tests the results are much more consistent and reliable. I've tested extremes, like 3:1, 1:3, 4:1, 1:4, and everything seems even now. Tested different depth setings and FOV, and it all seems right to me now.