Branding Typo for Deform Sable Diffuson
This has been with “Deforum” since the original notebook where the keys were added. I knew about it, but it used to only be referenced in a few places. But now? OMG. I was shocked at how many I had to change.  How could it get so bad.

I MEAN, IT'S OUR NAME

Someone needed to do it.