Disabled optical flow cadence in 2D mode
I'll have to do some more work to track down a method to warp the flow properly in 2D mode. For now, optical flow cadence is a 3D only thing.

Made it say Optical flow cadence (3D only)
Made it hidden by default
Made it show when you switch to 3D mode
Made it not do the routines on the back end if not in 3D mode