Fix for optical flow cadence on strength 0
I wasn't thinking about it quite right before, and I was cancelling the wrong cadence section based on strength 0. This repairs it.

It now looks at the proper strength key (from the tween frame_start_idx) to see if strength is 0 and skip optical flow cadence if it is.

Also, fixes and condenses the reporting of whether it's a cadence frame or an optical flow cadence by combining the info with the tween reporting.