Removed extra color match code for resume
This was an afterthought, and after further testing, I don't think it's working quite right. I'm not sure why, but removing for now. Will re-examine later.

I accidentally had color matching against video on instead of LAB at first when I was testing, so I didn't see it.  Maybe I just need to convert to BGR?

Anyway, this restores the previous behavior of color matching, where if it resumes, it just matches against the prev_img.