quick fix for progress bar during cadence
makes progress advance smoothly during cadence rather than skipping ahead suddenly after cadence completes

left line commented for state.current_image, since I'm not sure what that variable actually does. It doesn't update the preview when I try it, even when turning up the speed of preview updates in auto1111.  But, if we decide to make a preview mechanism or need that state var for cadence, it can just be uncommented.  It matches the format of the one in the main non-cadence loop.