updated transform_image_3d_new
removed excess options, removed dev code, cleaned up code, ready for production.
- removed depth tensor autocontrast options
- made equalization a part of the normalization process (not an optional arg now)
- removed my mechanism for 0-based normalization. Decided it wasn't any better. Maybe worse... mostly similar.
- cracked open a bottle of champagne on this function's hull

also added a file for the first steps of making an auto-navigation module to be used in the future inside the transform_image_3d function!  It generates a rotation matrix based on greatest or least depth in the tensor, instead of using animation keys.