Make gif output conditionally based on the interpolated frames.
This is to fix a bug where make_gif is checked and RIFE/FILM is applied. The output gif should be based off of the RIFE/FILM output.

Without this change, the make_gif will generate a gif from the non interpolated frames which defeats the purpose of RIFE/FILM if a user wants a gif output with highquality with gifski.